Cancel old animations when changing speed
Previously, we'd call toggleDraw() to pause and unpause the animation,
but draw() doesn't cancel itself when run so by the next frame it is
unpaused so the old draw() will not cancel.  Then we have created an
additional instance of draw() which causes issues.  Instead, store the
result of requestAnimationFrame and use it to cancelAnimationFrame when
we pause.