Cancel previous draw operations when new one starts
tripData and currentlyDrawing were both globals and they were both
updated when drawTrains() is called and therefore no drawings were
actually cancelled.  Fix this by making tripData local again.

draw was also a global, which is an issue because the newest function
would always get scheduled with requestAnimationFrame.  Make it tlocal
so we are scheduling the correct instance.

When loading new data and drawing in a paused state, things would break
because we call draw() while the code is paused.  Instead, call
toggleDraw() twice which will draw the new frame then pause.